chatbot/requirements.txt
pandas
langchain 
chromadb
groq
langchain-groq
PyPDF2
langchain_google_genai
streamlit
langchain_community
python-dotenv
google-generativeai
pymupdf
tiktoken 
langchainhub
langchain_huggingface
pillow 
transformers
streamlit
cassandra-driver
cassio
sentence-transformers
open-clip-torch
pandasai